Refatorando criação do parser para usar em outros testes
